Add files via upload
Traditionally, genre classification models are only based upon the audio features of songs. In this project, we seek to examine the impact of incorporating lyrics into these models, and whether or not it yields a greater accuracy in classification. We evaluate two approaches: a lyric-only model using BERT embeddings in a multiclass logistic regression, and a multimodal model where acoustic features are appended to BERT embeddings and fed into a multiclass logistic regression. Our experimental results show that the multimodal model outperforms the lyric-only model, with both approaches achieving significantly better results when compared to a majority classifier baseline.